# api/openapi.yaml
openapi: 3.0.3
info:
  title: HeartFolio API
  description: |
    HeartFolio transforms emotional pain and healing processes into verifiable soft-skill assets.

    ## JSON:API

    All endpoints conform to the [JSON:API v1.0](https://jsonapi.org/) specification.

    **Content-Type:** `application/vnd.api+json`

    ### Response envelope

    Every response is wrapped in a top-level `data` object (single resource) or `data` array (collection).

    ```json
    { "data": { "type": "users", "id": "abc123", "attributes": { ... } } }
    ```

    ### Error envelope

    All errors return an `errors` array instead of `data`.

    ```json
    { "errors": [{ "status": "404", "title": "Not Found", "detail": "badge not found" }] }
    ```

    ### Relationships

    Resources that embed related data include a `relationships` key whose value is a full
    resource object (not just a linkage), making it a single-request pattern.

    ### Authentication

    Protected endpoints require `Authorization: Bearer <token>`.
    Obtain a token from `POST /api/v1/users/login` or `POST /api/v1/users/register`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable schemas ──────────────────────────────────────────────────────
  schemas:
...
    ErrorObject:
      type: object
      properties:
        status:
          type: string
          example: "404"
        title:
          type: string
          example: "Not Found"
        detail:
          type: string
          example: "badge not found"
    ErrorDocument:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorObject"

    # ── Attribute bags ────────────────────────────────────────────────
    UserAttributes:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        exp:
          type: integer
          format: int64
          example: 250
        level:
          type: integer
          format: int32
          example: 3
    AuthAttributes:
      allOf:
        - $ref: "#/components/schemas/UserAttributes"
        - type: object
          properties:
            token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIs..."
...
    UserResource:
      type: object
      properties:
        type:
          type: string
          enum: [users]
        id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        attributes:
          $ref: "#/components/schemas/UserAttributes"
    AuthResource:
      type: object
      properties:
        type:
          type: string
          enum: [users]
        id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        attributes:
          $ref: "#/components/schemas/AuthAttributes"
...
    UserDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/UserResource"
    AuthDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/AuthResource"
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"
          example:
            errors:
              - status: "401"
                title: "Unauthorized"
                detail: "missing authorization header"
...
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/users/register:
    post:
      tags: [Users]
      summary: Register a new user
      description: Creates a new account and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "secret123"
      responses:
        "201":
          description: User created
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/AuthDocument"
              example:
                data:
                  type: users
                  id: "665f1a2b3c4d5e6f7a8b9c0d"
                  attributes:
                    email: "user@example.com"
                    role: "user"
                    token: "eyJhbGciOiJIUzI1NiIs..."
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: User already exists
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/ErrorDocument"
  /api/v1/users/login:
    post:
      tags: [Users]
      summary: Login
      description: Validates credentials and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: Login successful
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/AuthDocument"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserDocument"
              example:
                data:
                  type: users
                  id: "665f1a2b3c4d5e6f7a8b9c0d"
                  attributes:
                    email: "user@example.com"
                    role: "user"
                    exp: 250
                    level: 3
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Badges (public) ────────────────────────────────────────────────────────
...